fix some module dependencies
