+++ src/client/app/components/home/home.module.ts
@@ -3,9 +3,11 @@ import {CommonModule} from '@angular/common';
 import {MainModule} from '../../frameworks/ngModules/main.module';
 import {HomeComponent} from './home.component';
 import {NameListService} from '../../frameworks/app/index';
+import {FormsModule} from '@angular/forms';
+import {TranslateModule} from 'ng2-translate';
 
 @NgModule({
-    imports: [CommonModule, MainModule],
+    imports: [CommonModule, MainModule, TranslateModule.forRoot(), FormsModule],
     declarations: [HomeComponent],
     exports: [HomeComponent],
     providers: [NameListService]
