+++ src/client/app/frameworks/ngModules/main.module.ts
@@ -1,21 +1,27 @@
 import { NgModule, ModuleWithProviders } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { FormsModule } from '@angular/forms';
 import { RouterModule } from '@angular/router';
 
 import { ToolbarComponent } from '../../components/app/toolbar.component';
 import { NavbarComponent } from '../../components/app/navbar.component';
 import { NameListService } from '../app/index';
+import { TranslateModule } from 'ng2-translate';
+import {FormsModule} from '@angular/forms';
 
 /**
  * Do not specify providers for modules that might be imported by a lazy loaded module.
  */
 
 @NgModule({
-  imports: [CommonModule, RouterModule],
+  imports: [
+    CommonModule,
+    RouterModule,
+    TranslateModule.forRoot(),
+    FormsModule
+  ],
   declarations: [ToolbarComponent, NavbarComponent],
   exports: [ToolbarComponent, NavbarComponent,
-    CommonModule, FormsModule, RouterModule]
+    CommonModule, RouterModule]
 })
 export class MainModule {
   static forRoot(): ModuleWithProviders {
